Tweaked link to gif in README after migrating this repo's location
This repo used to be a private repository hosted on the github account
that my school provisioned me. I've since moved it over to my personal,
public github account, so I needed to have the gif on the README point
to the version of that graphic in this new repo.